Refactor (again) for E & D classes
