+++ python/baseline/pytorch/transformer.py
@@ -4,8 +4,8 @@ import torch.nn.functional as F
 import math
 import numpy as np
 from baseline.pytorch.torchy import pytorch_linear, pytorch_activation, LayerNorm
-from baseline.pytorch.torchy import pytorch_clone_module
-
+from baseline.pytorch.torchy import pytorch_clone_module, sequence_mask
+from collections import namedtuple
 
 
 def subsequent_mask(size):
